+++ src/main/java/cz/cuni/lf1/lge/ThunderSTORM/estimators/PSF/EllipticGaussianPSF.java
@@ -74,7 +74,7 @@ public class EllipticGaussianPSF extends PSFModel {
     }
 
     @Override
-    public MultivariateMatrixFunction getJacobianFunction(final int[] xgrid, final int[] ygrid) {
+    public MultivariateMatrixFunction getJacobianFunction(final double[] xgrid, final double[] ygrid) {
         return new MultivariateMatrixFunction() {
             @Override
             public double[][] value(double[] point) throws IllegalArgumentException {
