+++ src/main/java/cz/cuni/lf1/lge/ThunderSTORM/estimators/PSF/SymmetricGaussianPSF.java
@@ -54,7 +54,7 @@ public class SymmetricGaussianPSF extends PSFModel {
     }
 
     @Override
-    public MultivariateMatrixFunction getJacobianFunction(final int[] xgrid, final int[] ygrid) {
+    public MultivariateMatrixFunction getJacobianFunction(final double[] xgrid, final double[] ygrid) {
         return new MultivariateMatrixFunction() {
             @Override
             //derivations by wolfram alpha:
